support array flattening for json
